% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSynthesis_advanced.R
\name{ABSynthesis_advanced}
\alias{ABSynthesis_advanced}
\title{Combine AB test results and get advanced statistics for each study in the set}
\usage{
ABSynthesis_advanced(successes_base, base_n, successes_variant, variant_n)
}
\arguments{
\item{successes_base}{Count of responses in the base condition}

\item{base_n}{Count of visitors exposed to the base condition}

\item{successes_variant}{Count of responses in the base condition}

\item{variant_n}{Count of visitors exposed to the base condition}
}
\value{
Summary statistics of a meta-analysis of online controlled experiments
}
\description{
ABSynthesis_advanced returns results of each study in the set accounting for the overall effect of all other studies.
}
\examples{
ABSynthesis_advanced(c(10, 10), c(50, 50), c(12, 12), c(50, 50))
}
