% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rapidapi_key.R
\name{add_rapidapi_key}
\alias{add_rapidapi_key}
\title{Add a RapidAPI key to your .Renviron}
\usage{
add_rapidapi_key(key, renviron_path = "~/.Renviron")
}
\arguments{
\item{key}{The number of visitors that should be simulated}

\item{renviron_path}{Whether the study should include return visitors}
}
\value{
Updates the .Renviron with the key
}
\description{
add_rapidapi_key is a helper function that adds your RapidAPI key to your R environment. You will need to provide an API key which can be downloaded from rapidapi.com/hub .
}
