% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Get a Forest plot of combined experiments}
\usage{
forest_plot(
  experiment_ids = NA,
  successes_base,
  visitors_base,
  successes_variant,
  visitors_variant
)
}
\arguments{
\item{experiment_ids}{A list of unique experiment ids if available}

\item{successes_base}{Count of responses in the base condition}

\item{visitors_base}{Count of visitors exposed to the base condition}

\item{successes_variant}{Count of responses in the base condition}

\item{visitors_variant}{Count of visitors exposed to the base condition}
}
\value{
A forest plot stacking the results of multiple tests of the same element, page, site, app, etc.
}
\description{
forest_plot returns summary statistics combining the results of all online controlled experiments
}
\examples{
forest_plot(c(1,2,3), c(10, 15,20), c(50, 50, 50), c(12, 17, 22), c(50, 50, 50))
}
