% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_AB_corpus.R
\name{initialize_synthetic_AB_corpus}
\alias{initialize_synthetic_AB_corpus}
\title{Generate a set of synthetic experiment results based on desired conversion probabilities}
\usage{
initialize_synthetic_AB_corpus(
  experiments,
  minimum_visitors,
  maximum_visitors,
  baseline_probability,
  variant_probability
)
}
\arguments{
\item{experiments}{The number of experiments that should be generated}

\item{minimum_visitors}{The number of visitors that should be simulated}

\item{maximum_visitors}{The number of visitors that should be simulated}

\item{baseline_probability}{Proportion of visitors that perform the desired action in the base version}

\item{variant_probability}{Proportion of visitors that perform the desired action in the variant version}
}
\value{
A set of numbers that represent what would be seen at the end of one AB test with the underlying input parameters.
}
\description{
initialize_synthetic_AB_corpus creates a set of numbers that represent what may be seen at the end of a set of AB tests.
}
\examples{
initialize_synthetic_AB_corpus(50, 5000,10000, 0.1, 0.2)
}
